+++ packages/utils/telemetry-utils/src/config.ts
@@ -167,6 +167,24 @@ const safeSessionStorage = (): Storage | undefined => {
 };
 
 /**
+ * Creates a wrapper on top of an existing config provider which allows for
+ * specifying feature gates if not present in the original provider.
+ *
+ * @param original - the original config provider
+ * @param defaults - default feature gate configs to be used if not specified by the original provider
+ * @returns A config provider that looks for any requested feature gates in the original provider and falls
+ * back to the values specified in the `defaults` feature gates if they're not present in the original.
+ *
+ * @internal
+ */
+export const wrapConfigProviderWithDefaults = (
+	original: IConfigProviderBase | undefined,
+	defaults: Record<string, ConfigTypes>,
+): IConfigProviderBase => ({
+	getRawConfig: (name: string): ConfigTypes => original?.getRawConfig(name) ?? defaults[name],
+});
+
+/**
  * Implementation of {@link IConfigProvider} which contains nested {@link IConfigProviderBase} instances
  */
 export class CachedConfigProvider implements IConfigProvider {