+++ examples/apps/presence-tracker/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @fluid-example/presence-tracker
 
+## 2.0.0-rc.1.0.0
+
+Dependency updates only.
+
 ## 2.0.0-internal.8.0.0
 
 Dependency updates only.