+++ examples/version-migration/live-schema-upgrade/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @fluid-example/app-integration-live-schema-upgrade
 
+## 2.0.0-rc.1.0.0
+
+Dependency updates only.
+
 ## 2.0.0-internal.8.0.0
 
 Dependency updates only.